// Tiket/src/app/Models/hotel.ts
export class Hotel {
    id: number;
    name: string;
    lat: number;
    long: number;
    rating: number;
    province: string;
    city: string;
    address: string;
    path: string;
    facility: string;
    desc: string;
    price: number;
    type: string;
    constructor(){
        
    }
}